added logic in addComment component
